service_wrapper_project/requirements/base.txt
django>=1.11.5
django-environ>=0.4.4
djangorestframework>=3.6.4
python3-indy
